backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.1.3
black==26.1.0
boto3==1.42.57
botocore==1.42.57
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
# ChromaDB uses OpenAI embeddings (openai package); we do not use onnxruntime.
chromadb>=0.4.0
# Pin to range required by google-adk; chroma_activities stubs ReadableLogRecord for compatibility
opentelemetry-api>=1.36.0,<1.39.0
opentelemetry-sdk>=1.36.0,<1.39.0
# Keep OTLP exporter in same range so it matches sdk (avoids proto-grpc 1.39 vs sdk 1.38 conflict)
opentelemetry-exporter-otlp-proto-grpc>=1.36.0,<1.39.0
# Opik tracing (OTLP over HTTP; Opik does not support gRPC)
opentelemetry-exporter-otlp-proto-http>=1.36.0,<1.39.0
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.133.1
google-adk==1.25.1
fastuuid==0.14.0
filelock==3.24.3
flake8==7.3.0
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==8.0.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.8.1
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
# Used by ChromaDB OpenAI embedding function for semantic search (OPENAI_API_KEY).
openai==1.99.9
# Trace OpenAI API calls (embeddings, etc.) in Opik; requires otel-api >=1.38
opentelemetry-instrumentation-openai>=0.46.0
packaging==26.0
pandas==3.0.1
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.5.0
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
sendgrid==6.12.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
stripe==14.4.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.6
yarl==1.22.0
zipp==3.23.0
authlib
